:bug: Fix issue with long input strings to the Python interpreter
This refactors code execution to eliminate the input line from the
python process output stream and then add it back if echo is turned
on.  Before this, a long input string could get echoed back from
the interpreter with `\r` and repeated characters interspersed, and
this would corrupt the overall output with or without echo (to the
document) enabled.